requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dgl-cu110==0.6.1
googledrivedownloader==0.4
idna==2.10
importlib-metadata==4.6.1
iniconfig==1.1.1
isodate==0.6.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.0
packaging==21.0
pandas==1.2.5
Pillow==8.3.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-louvain==0.15
pytz==2021.1
rdflib==5.0.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.2
six==1.16.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.0+cu111
torch-cluster==1.5.9
torch-geometric==1.7.2
torch-scatter==2.0.8
torch-sparse==0.6.11
torch-spline-conv==1.2.1
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wincertstore==0.2
zipp==3.5.0
